code/requirements.txt
absl-py==1.4.0
addict==2.4.0
-e git+https://github.com/BorealisAI/advertorch.git@e063e341c87d9a621ae1a8f72c3507d5ea5bd327#egg=advertorch
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1694046349000/work
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autograd==1.6.2
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
beautifulsoup4==4.12.2
black==21.4b2
cachetools==5.3.1
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
cityscapesScripts==2.2.2
click==8.1.4
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.6
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.1.0
cPython==0.0.6
cycler==0.11.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
detectron2==0.6+cu113
dnspython==2.4.2
eagerpy==0.30.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
foolbox==3.3.3
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.10.0
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.3.2
idna==2.10
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1693880262622/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
joblib==1.2.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1696972407098/work
keras==2.13.1
kiwisolver==1.4.4
kornia==0.7.0
lazy_loader==0.2
libclang==16.0.6
lightning-utilities==0.9.0
lit==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mmcv==2.0.1
mmengine==0.8.1
model-index==0.1.11
mpmath==1.3.0
mtcnn-opencv==1.0.2
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.1
nose==1.3.7
numpy==1.20.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opendatalab==0.0.9
openmim==0.3.9
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.11.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1696272223550/work
portalocker==2.7.0
pretrainedmodels==0.7.4
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.23.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycryptodome==3.18.0
pyDeprecate==0.3.1
pydot==1.4.2
Pygments==2.15.1
pymongo==4.5.0
pyparsing==3.1.0
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.5.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq @ file:///croot/pyzmq_1686601365461/work
regex==2023.8.8
requests==2.25.1
requests-oauthlib==1.3.1
retina-face==0.0.13
rich==13.4.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==0.22
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
some-package==0.1
soupsieve==2.5
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.5
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchattacks==3.4.0
torchcam==0.4.0
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1696377679271/work
triton==2.0.0
typing==3.10.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1696255154857/work
Werkzeug==2.3.6
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.15.0